requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.11.0
openai==1.3.7
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
icalendar==5.0.5
websockets==12.0
